Conflicts:
- FlatList.tsx
- TempSchedShiftsList.tsx

+++ lock/global.go
@@ -2,10 +2,10 @@ package lock
 
 // Defined global lock values.
 const (
-	GlobalMigrate            = 0x1337 // 4919
-	GlobalEngineProcessing   = 0x1234 // 4660
-	GlobalMessageSending     = 0x1330 // 4912
-	RegionalEngineProcessing = 0x1342 // 4930
-	ModularEngineProcessing  = 0x1347 // 4935
-	GlobalSwitchOver         = 0x1111 // 4369
+	GlobalMigrate            = uint32(0x1337) // 4919
+	GlobalEngineProcessing   = uint32(0x1234) // 4660
+	GlobalMessageSending     = uint32(0x1330) // 4912
+	RegionalEngineProcessing = uint32(0x1342) // 4930
+	ModularEngineProcessing  = uint32(0x1347) // 4935
+	GlobalSwitchOver         = uint32(0x1111) // 4369
 )
